app/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pydantic>=2.6.0
# Optional — for richer file type support
pdfplumber>=0.10.0
python-docx>=1.1.0
